// frontend/src/app/core/api/model/accredited-promotion-response.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AccreditedPromotionResponse { 
    id?: number;
    name?: string;
    code?: string;
    startDate?: string;
    endDate?: string;
    status?: AccreditedPromotionResponse.StatusEnum;
    fee?: number;
    trainingId?: number;
    trainingLabel?: string;
    specialityLabel?: string;
    enrollmentCount?: number;
    registrationOpeningDate?: string;
    registrationDeadline?: string;
    academicYearId?: number;
    academicYearLabel?: string;
}
export namespace AccreditedPromotionResponse {
    export const StatusEnum = {
        Planned: 'PLANNED',
        EnrollmentOpen: 'ENROLLMENT_OPEN',
        InProgress: 'IN_PROGRESS',
        Closed: 'CLOSED'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
}